Finish up clang-tidy

+++ modyn/NewStorage/include/internal/file_watcher/file_watchdog.hpp
@@ -17,7 +17,8 @@ class FileWatchdog {
  private:
   YAML::Node config_;
   std::string config_file_;
-  std::unordered_map<int64_t, std::tuple<std::thread, int, std::shared_ptr<std::atomic<bool>>>> file_watcher_processes_;
+  std::unordered_map<int64_t, std::tuple<std::thread, int16_t, std::shared_ptr<std::atomic<bool>>>>
+      file_watcher_processes_;
   std::shared_ptr<std::atomic<bool>> stop_file_watchdog_;
 
  public:
@@ -25,10 +26,10 @@ class FileWatchdog {
       : config_file_(config_file), stop_file_watchdog_(std::move(stop_file_watchdog)) {
     config_ = YAML::LoadFile(config_file);
     file_watcher_processes_ =
-        std::unordered_map<int64_t, std::tuple<std::thread, int, std::shared_ptr<std::atomic<bool>>>>();
+        std::unordered_map<int64_t, std::tuple<std::thread, int16_t, std::shared_ptr<std::atomic<bool>>>>();
   }
   void watch_file_watcher_processes(StorageDatabaseConnection* storage_database_connection);
-  void start_file_watcher_process(int64_t dataset_id, int retries);
+  void start_file_watcher_process(int64_t dataset_id, int16_t retries);
   void stop_file_watcher_process(int64_t dataset_id, bool is_test = false);
   void run();
   std::vector<int64_t> get_running_file_watcher_processes();
