+++ modyn/NewStorage/test/unit/internal/file_wrapper/mock_file_wrapper.hpp
@@ -12,12 +12,12 @@ namespace storage {
 class MockFileWrapper : public AbstractFileWrapper {
  public:
   MockFileWrapper() : AbstractFileWrapper("", YAML::Node(), nullptr){};
-  MOCK_METHOD(int, get_number_of_samples, (), (override));
-  MOCK_METHOD(std::vector<std::vector<unsigned char>>*, get_samples, (int start, int end), (override));
-  MOCK_METHOD(int, get_label, (int index), (override));
-  MOCK_METHOD(std::vector<int>*, get_all_labels, (), (override));
-  MOCK_METHOD(std::vector<unsigned char>*, get_sample, (int index), (override));
-  MOCK_METHOD(std::vector<std::vector<unsigned char>>*, get_samples_from_indices, (std::vector<int> * indices),
+  MOCK_METHOD(int64_t, get_number_of_samples, (), (override));
+  MOCK_METHOD(std::vector<std::vector<unsigned char>>*, get_samples, (int64_t start, int64_t end), (override));
+  MOCK_METHOD(int64_t, get_label, (int64_t index), (override));
+  MOCK_METHOD(std::vector<int32_t>*, get_all_labels, (), (override));
+  MOCK_METHOD(std::vector<unsigned char>*, get_sample, (int64_t index), (override));
+  MOCK_METHOD(std::vector<std::vector<unsigned char>>*, get_samples_from_indices, (std::vector<int64_t> * indices),
               (override));
   MOCK_METHOD(std::string, get_name, (), (override));
   MOCK_METHOD(void, validate_file_extension, (), (override));
